% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayMAP_package.R
\docType{package}
\name{BayMAP-package}
\alias{BayMAP}
\alias{BayMAP-package}
\title{BayMAP: Bayesian Hierarchical Model for the Analysis of PAR-CLIP data}
\description{
Running BayMAP on PAR-CLIP data to detect PAR-CLIP induced T-to-C substitution positions on binding sites.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/csdaw/BayMAP}
  \item Report bugs at \url{https://github.com/csdaw/BayMAP/issues}
}

}
\author{
\strong{Maintainer}: Charlotte Dawson \email{csdaw@outlook.com} (\href{https://orcid.org/0000-0002-7151-5971}{ORCID}) [contributor]

Authors:
\itemize{
  \item Eva-Maria Huessler \email{eva-maria.huessler@uni-duesseldorf.de}
}

}
\keyword{internal}
